+++ src/backend/base/langflow/services/telemetry/schema.py
@@ -1,0 +1,35 @@
+from pydantic import BaseModel
+
+
+class RunPayload(BaseModel):
+    IsWebhook: bool = False
+    seconds: int
+    success: bool
+    errorMessage: str = ""
+
+
+class ShutdownPayload(BaseModel):
+    timeRunning: int
+
+
+class VersionPayload(BaseModel):
+    version: str
+    platform: str
+    python: str
+    arch: str
+    autoLogin: bool
+    cacheType: str
+    backendOnly: bool
+
+
+class PlaygroundPayload(BaseModel):
+    seconds: int
+    componentCount: int
+    success: bool
+
+
+class ComponentPayload(BaseModel):
+    name: str
+    seconds: int
+    success: bool
+    errorMessage: str
